Activity CRUD hecho falta los listado de actividades de un día y de una semana

+++ symfony/src/Category/Application/UseCase/DeleteCategory/DeleteCategoryHandler.php
@@ -2,6 +2,7 @@
 
 namespace App\Category\Application\UseCase\DeleteCategory;
 
+use App\Category\Domain\Exception\CategoryNotFoundException;
 use App\Category\Domain\Exception\CategoryNotOwnedByUserException;
 use App\Category\Domain\Repository\CategoryRepositoryInterface;
 use App\Category\Domain\ValueObject\CategoryId;
@@ -14,10 +15,16 @@ class DeleteCategoryHandler
 
     /**
      * @throws \App\Category\Domain\Exception\CategoryNotOwnedByUserException
+     * @throws \App\Category\Domain\Exception\CategoryNotFoundException
      */
     public function __invoke(DeleteCategoryCommand $command): void
     {
         $category = $this->categoryRepository->findById(new CategoryId($command->categoryId));
+
+        if (!$category) {
+            throw new CategoryNotFoundException();
+        }
+
         if ($category->getUserId()->getUuid() == $command->userId) {
             $this->categoryRepository->delete($category->getId());
         } else {
